ai-service/requirements.txt
tensorflow==2.12.0
kafka-python==2.0.2
numpy==1.23.5
pandas==2.0.1
scikit-learn==1.2.2
python-dotenv==1.0.0
fastapi==0.68.0
uvicorn==0.15.0
dnspython==2.1.0
python-whois==0.7.3
requests==2.26.0
pydantic==1.8.2
loguru==0.7.0